사용자 후기 어플리케이션 스테이지 설정 및 실행 아이콘 작업

+++ src/main/java/review/impl/ReviewServiceImpl.java
@@ -11,7 +11,7 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public List<Review> selectMemberReview(long memberId) throws Exception {
-        List<Review> reviews = reviewDao.selectMemberReview((int) memberId);
+        List<Review> reviews = reviewDao.selectMemberReview(memberId);
 
         for (Review review : reviews) {
             System.out.println("Review ID: " + review.getId());
